Unify move shapes in a util method
