+++ libs/statemanager/src/main/kotlin/mono/state/command/mouse/MoveShapeMouseCommand.kt
@@ -7,9 +7,7 @@ package mono.state.command.mouse
 import mono.common.MouseCursor
 import mono.graphics.geo.MousePointer
 import mono.graphics.geo.Point
-import mono.shape.command.ChangeBound
 import mono.shape.shape.AbstractShape
-import mono.shape.shape.Line
 import mono.state.command.CommandEnvironment
 import mono.state.command.mouse.MouseCommand.CommandResultType
 import mono.state.utils.UpdateShapeBoundHelper
@@ -36,35 +34,11 @@ internal class MoveShapeMouseCommand(private val shapes: Set<AbstractShape>) : M
             MousePointer.Idle -> Point.ZERO
         }
 
-        val isUpdateConfirmed = mousePointer is MousePointer.Up
-
-        val affectedLines = mutableSetOf<String>()
-        val (lines, notLineShapes) = shapes.partition { it is Line }
-
-        for (shape in notLineShapes) {
-            val initialPosition = initialPositions[shape.id] ?: continue
-            val newPosition = initialPosition + offset
-            val newBound = shape.bound.copy(position = newPosition)
-
-            environment.shapeManager.execute(ChangeBound(shape, newBound))
-            affectedLines += UpdateShapeBoundHelper.updateConnectors(
-                environment,
-                shape,
-                newBound,
-                isUpdateConfirmed
-            )
-        }
-
-        for (line in lines) {
-            if (line.id in affectedLines) {
-                // If a line is updated by connectors, ignore it from updating as individual shape
-                continue
-            }
-            val initialPosition = initialPositions[line.id] ?: continue
-            val newPosition = initialPosition + offset
-            val newBound = line.bound.copy(position = newPosition)
-            environment.shapeManager.execute(ChangeBound(line, newBound))
-        }
+        UpdateShapeBoundHelper.moveShapes(
+            environment,
+            shapes,
+            isUpdateConfirmed = mousePointer is MousePointer.Up
+        ) { shape -> initialPositions[shape.id]?.let { it + offset } }
 
         environment.updateInteractionBounds()
