+++ libs/statemanager/src/main/kotlin/mono/state/utils/UpdateShapeBoundHelper.kt
@@ -4,7 +4,9 @@
 
 package mono.state.utils
 
+import mono.graphics.geo.Point
 import mono.graphics.geo.Rect
+import mono.shape.command.ChangeBound
 import mono.shape.connector.ShapeConnectorUseCase.getPointInNewBound
 import mono.shape.shape.AbstractShape
 import mono.shape.shape.Line
@@ -14,6 +16,34 @@ import mono.state.command.CommandEnvironment
  * A utility object for updating connectors attached to a shape.
  */
 internal object UpdateShapeBoundHelper {
+    fun moveShapes(
+        environment: CommandEnvironment,
+        selectedShapes: Collection<AbstractShape>,
+        isUpdateConfirmed: Boolean,
+        newPositionCalculator: (AbstractShape) -> Point?
+    ) {
+        val affectedLines = mutableSetOf<String>()
+        val (lines, notLineShapes) = selectedShapes.partition { it is Line }
+
+        for (shape in notLineShapes) {
+            val newPosition = newPositionCalculator(shape) ?: continue
+            val newBound = shape.bound.copy(position = newPosition)
+
+            environment.shapeManager.execute(ChangeBound(shape, newBound))
+            affectedLines += updateConnectors(environment, shape, newBound, isUpdateConfirmed)
+        }
+
+        for (line in lines) {
+            if (line.id in affectedLines) {
+                // If a line is updated by connectors, ignore it from updating as individual shape
+                continue
+            }
+            val newPosition = newPositionCalculator(line) ?: continue
+            val newBound = line.bound.copy(position = newPosition)
+            environment.shapeManager.execute(ChangeBound(line, newBound))
+        }
+    }
+
     /**
      * Updates connects of [target].
      * Returns a list of affected lines
